+++ src/seamlesserverpkg/renderer/base/footer.nim
@@ -1,0 +1,6 @@
+include pkg/karax/prelude
+
+proc renderFooter*: VNode =
+  ## Render page footer
+  buildHtml(footer):
+    text "footer"