+++ src/seamlesserverpkg/renderer/base/header.nim
@@ -1,0 +1,6 @@
+include pkg/karax/prelude
+
+proc renderHeader*: VNode =
+  ## Render page header
+  buildHtml(header):
+    text "header"