+++ changelog.md
@@ -5,6 +5,9 @@
 - Added SPA routing without reloading
 - Added "flash" errors
 - Added BridgedData, data which is provided by backend, directly to frontend
+- Removed useless extra tdiv at frontend rendering
+- Added header and footer
+- Added blank register page
 
 ## Version alpha 0.0.1 (26/06/2023)